tests/test_auth.py
import pytest
from flask import url_for
from flask_jwt_extended import decode_token

def test_login_jwt(client):
    # Replace with valid test user credentials
    response = client.post('/login', data={'email': 'test@user.com', 'password': 'testpass'})
    assert response.status_code == 200 or response.status_code == 302
    # Check for access token in cookies or response
    assert 'access_token' in response.headers.get('Set-Cookie', '')

def test_role_access(client):
    # Try to access admin route without token
    response = client.get('/admin/dashboard')
    assert response.status_code in (302, 401, 403)

def test_2fa_flow(client):
    # Simulate login for doctor/admin and OTP step (pseudo)
    # You must implement a real OTP test with a known secret
    pass  # Placeholder
