requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1 
pyotp==2.9.0
qrcode==7.4.2
cryptography==41.0.3
google-auth-oauthlib==1.2.0
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
google-auth==2.40.1
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pypng==0.20220715.0
rsa==4.9.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3